convert ObtainResources to .sbt format
and simplify it a lot now that we're mostly getting stuff from webjars
not the ccl server

(also make the task be a Task[Seq[File]] not a Task[Unit]. we're not
currently using the return value, but this is just generally the
better way to do it. e.g. it makes `show obtainStylesheets` work)