Feature: Display non-compiling models as disabled
Originally, we showed red Xs and green checks, based on the compilation
status.  Some people found that confusing, so we just decided not to
show people models in Production that they can't actually use.  Uri,
however, feels that it is important that users be shown the whole Models
Library, so we are now disabling selection of broken models, instead.