Refactor: Store Beak widgets in an object (not array)
This makes deleting widgets easier.  If the widgets are in an array, and
you delete one, you either need to create a sparse array (which is
annoying, and trips up the compiler, and needs to be worked around by
having consumers always filter out the `undefined`s), or you run into
issues with Ractive getting confused about which index in the array
to be grabbing the correct widget from (which I don't know how to
work around).

Seems to me like the smarter thing to do is just to put the widgets into
a structure that is made for absolute identifiers and removing elements
from the middle mid-run (i.e. objects).