Bugfix/UX: Size accordion more reasonably
Ultimately, this corrects a regression from ae5547a.  The Code pane
no longer scrolled horizontally, and it would just take as much space
as it needed to fit everything in without scrolling.  This led to some
confusing behavior, with things not being visible where we'd expect to
see them, since they were horizontally centered in a massively wide
accordion.