Bugfix: Don't create multiple sessions for uploaded model
Fixes #372

I accidentally introduced this in 3b0753c.  As a result, everytime a
model was uploaded—and only with uploads—*two new* sessions would be
created.  As it turns out, this was even breaking drawings for uploaded
models, since they were going to the invisible canvas of one of those
two new sessions (but the wrong one).  But, still... we went from having
3 sessions at once to having 2 sessions at once.  But that's still one
too many!  I'm going to try to figure out why it's leaking the prior
sessions for uploads soon.