Infrastructure: Remove autoprefixer
The `sbt-autoprefixer` plugin continues to cause problems for building and deploying NetLogo Web, and after discussing with the team we decided it may no longer be useful for us given that several years worth of browser CSS handling improvements have been made.  I removed it and tested the latest Firefox, Chrome, and Safari all on macOS High Sierra, and also tested Mobile Safari on iOS 11.2 in the Xcode iPad simulator.  Basic funcationality as well as a suite of models were tested, and I could not find any issues.