Refactor: Rethink drawing event management
Alternatively: "Let's *add* some drawing event management!"?

Previously, what we were doing was immediately applying all of our
drawing events and then throwing the updates away.  In some regard this
is sensible, as a model with a ton of pen/stamp events is going to
accumulate a large drawing event list.  If this becomes a problem, we
might need to do some periodic snapshotting of the View to condense all
of the existing events to a single "import-drawing" event.

Regardless, the old way was not really cool.  Since we weren't keeping
track of our drawing events, we had nothing to really send to HubNet
clients when they connected.

That was not an enormous problem, and we could have worked around that
by asking the model for its drawing layer PNG upon connection, but I'd
rather not be special-casing drawing layer stuff.  We shouldn't assume
that all clients want our pre-rendered drawing layer PNGs, and we can
just leverage the same system that all of our other diffs use.