Bugfix: Recompile model on button forever type change
When a button is setup it sets a `wrappedTask` for its `run` value, but
that only happens on recompile.  If you swapped the `forever` type of a
button in authoring from false to true it would not get the correct
wrapped task set, and if it threw an error while running you'd hit an
infinite loop of errors as the `running` value would never get set to
false.