Fix wrong path of CodeMirror assets
Some paths to CodeMirror assets were pointing to '/assets/lib/codemirror', but the files are currently located in '/assets/codemirror'. As a result, CodeMirror didn't load on the 'differences' page.

This apparently didn't cause a problem on the production website, because there were still old files (last changed in 2016) in '/assets/lib/codemirror', so the library could load both from the old/wrong and the new/correct location.