Feature: Check insecure linked model protocol and recover/alert
If running the page in HTTPS and the linked model is HTTP, normally
you'd have just seen a generic connection error.  If the linked model is
on the same site as the NLW page running it, just swap it to HTTPS as
well.  If the linked model is on another site we don't control, link to
the possible HTTPS version for non-iframe pages, otherwise just provide
the user the appropriate info and possible working model link.