Updates to WIP feature in settings
I modified the WIP to accommodate storage tags, and adding them as a display. In addition, this commit filters disk and new prefixes. We check that the version in local storage is the latest from wip-data.coffee and skip versions that are older. Finally, added a sorting function to ensure that the the WIP is properly sorted based on time and date.