HNW/Bugfix: Determine widget drag coords better
There are actually two different fixes here.

On the one hand, the widget editor was entirely broken.
Firefox is feeding nonsense coordinates into the JavaScript,
and those coordinates were often above 0 (but static), and
things were not moving.  So one part of the fix here never trusts
coordinates in Firefox, and always looks around for the mouse coords
that were cached elsewhere.

The other problem was that, right before 'dragend', the final 'drag'
event would just have these nonsense values in Chrome.  There's always
at least one negative coordinate.  So, when we detect that we have a
nonsense coordinate, we go and grab the cached value instead.