Teardown color picker keyup Esc event properly
c288c2 introduced a bug where the teardown function would attempt to
remove a keydown (now keyup) event listener with the wrong reference,
preventing it from ever being removed. This commit fixes that by
correctly adding and removing the keyup event listener.

By the time this commit was made, the color picker iframe was modified
to send a cancel event on Esc keyup as well, which c288c2 could not handle
when iframe was focused.